refactor stopwatch into it's own component
