+++ src/App.tsx
@@ -1,24 +1,9 @@
-import { useState } from 'react'
+import Stopwatch from './stopwatch'
 import pomogattoLogo from './assets/pomogatto.jpg'
 
 import './App.css'
 
 function App() {
-  const [timer, setTimer] = useState(0)
-  const [intervalId, setIntervalId] = useState(0)
-
-  function startTimer() {
-    const id = setInterval(() => { setTimer((timer) => timer + 1) }, 1000)
-    setIntervalId(id)
-  }
-
-  function stopTimer() {
-    clearInterval(intervalId);
-  }
-
-  function clearTimer() {
-    setTimer(0)
-  }
 
   return (
     <>
@@ -28,20 +13,7 @@ function App() {
         </a>
       </div>
       <h1>pomogatto</h1>
-      <div className="card">
-        <p>time elapsed is: {timer}</p>
-        <button onClick={startTimer}>
-          start timer
-        </button>
-
-        <button onClick={stopTimer}>
-          stop timer
-        </button>
-
-        <button onClick={clearTimer}>
-          clear timer
-        </button>
-      </div>
+      <Stopwatch />
     </>
   )
 }
